add component MainCard and organize general tree
