+++ buon-appetito/components/Header.module.css
@@ -6,7 +6,6 @@
   justify-content: center;
   align-items: center;
   background-color: #F6F3F0;
-  border-bottom: solid 1px lightgray;
 }
 
 .logo {
