+++ buon-appetito/components/Layout.js
@@ -1,0 +1,36 @@
+import Head from "next/head";
+
+import Header from "./Header";
+import Footer from "./Footer";
+
+export default function Layout({ children }) {
+  return (
+    <>
+      <Head>
+        <title>Buon APPetito</title>
+        <meta name="description" content="Generated by create next app" />
+        <link
+          rel="apple-touch-icon"
+          sizes="180x180"
+          href="/favicon/apple-touch-icon.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="32x32"
+          href="/favicon/favicon-32x32.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="16x16"
+          href="/favicon/favicon-16x16.png"
+        />
+        <link rel="manifest" href="/favicon/site.webmanifest" />
+      </Head>
+      <Header />
+      {children}
+      <Footer />
+    </>
+  );
+}
