+++ buon-appetito/components/MainCard.module.css
@@ -1,0 +1,22 @@
+.card {
+  margin: 0.5rem;
+  padding: 1rem;
+  text-align: left;
+  color: inherit;
+  text-decoration: none;
+  background-color: #3A94461A;
+  border: 2px solid #3A9446;
+  border-radius: 8px;
+  max-width: 400px;
+}
+
+.card h2 {
+  margin: 0 0 1rem 0;
+  font-size: 1.5rem;
+}
+
+.card p {
+  margin: 0;
+  font-size: 1.25rem;
+  line-height: 1.5;
+}
