--- buon-appetito/components/layout.js
@@ -1,14 +1,0 @@
-import Header from "./header";
-import Footer from "./footer";
-
-import styles from "./layout.module.css";
-
-export default function Layout({ children }) {
-  return (
-    <>
-      <Header />
-      <main>{children}</main>
-      <Footer />
-    </>
-  );
-}
