+++ buon-appetito/pages/index.module.css
@@ -1,0 +1,8 @@
+.container {
+  height: 100vh;
+  padding: 100px 1rem 80px;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+}
